// apps/web/src/shared/api/schema.ts
export interface components {
    schemas: {
// ...
        Category: {
            /** Format: uuid */
            id: string;
            /** Format: uuid */
            userId: string;
            name: string;
            /** @enum {string} */
            type: "income" | "expense";
            icon: string;
            color: string;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
        CategoryCreateRequest: {
            name: string;
            /** @enum {string} */
            type: "income" | "expense";
            icon: string;
            color: string;
        };
        /** @description Все поля optional. */
        CategoryUpdateRequest: {
            name?: string;
            /** @enum {string} */
            type?: "income" | "expense";
            icon?: string;
            color?: string;
        };
    };
// ...
}